move multivector to gko:: namespace
