+++ core/CMakeLists.txt
@@ -40,6 +40,7 @@ target_sources(
         base/memory.cpp
         base/mpi.cpp
         base/mtx_io.cpp
+        base/multivector.cpp
         base/perturbation.cpp
         base/segmented_array.cpp
         base/timer.cpp
@@ -89,7 +90,6 @@ target_sources(
         matrix/fft.cpp
         matrix/hybrid.cpp
         matrix/identity.cpp
-        matrix/multivector.cpp
         matrix/permutation.cpp
         matrix/row_gatherer.cpp
         matrix/scaled_permutation.cpp
