+++ include/ginkgo/core/base/multivector.hpp
@@ -12,13 +12,15 @@
 #include "ginkgo/core/base/type_traits.hpp"
 
 
-
 namespace gko {
 namespace matrix {
+
 
 template <typename ValueType>
 class Dense;
 
+
+}
 
 using supported_value_types =
     std::tuple<double, float, std::complex<double>, std::complex<float>
@@ -32,7 +34,7 @@ using supported_value_types =
 #endif
                >;
 
-using dense_types = syn::apply_to_list<Dense, supported_value_types>;
+using dense_types = syn::apply_to_list<matrix::Dense, supported_value_types>;
 
 using any_const_dense_t = syn::variant_from_tuple<syn::apply_to_list<
     ptr_param, syn::apply_to_list<std::add_const_t, dense_types>>>;
@@ -358,46 +360,46 @@ protected:
 
     virtual void fill_impl(value_type value) = 0;
 
-    virtual void scale_impl(const Dense<value_type>* alpha) = 0;
+    virtual void scale_impl(const matrix::Dense<value_type>* alpha) = 0;
 
-    virtual void inv_scale_impl(const Dense<value_type>* alpha) = 0;
+    virtual void inv_scale_impl(const matrix::Dense<value_type>* alpha) = 0;
 
-    virtual void add_scaled_impl(const Dense<value_type>* alpha,
+    virtual void add_scaled_impl(const matrix::Dense<value_type>* alpha,
                                  const ConcreteType* b) = 0;
 
-    virtual void sub_scaled_impl(const Dense<value_type>* alpha,
+    virtual void sub_scaled_impl(const matrix::Dense<value_type>* alpha,
                                  const ConcreteType* b) = 0;
 
     virtual void compute_dot_impl(const ConcreteType* b,
-                                  Dense<value_type>* result) const = 0;
+                                  matrix::Dense<value_type>* result) const = 0;
 
     virtual void compute_dot_impl(const ConcreteType* b,
-                                  Dense<value_type>* result,
+                                  matrix::Dense<value_type>* result,
                                   array<char>& tmp) const = 0;
 
-    virtual void compute_conj_dot_impl(const ConcreteType* b,
-                                       Dense<value_type>* result) const = 0;
+    virtual void compute_conj_dot_impl(
+        const ConcreteType* b, matrix::Dense<value_type>* result) const = 0;
 
     virtual void compute_conj_dot_impl(const ConcreteType* b,
-                                       Dense<value_type>* result,
+                                       matrix::Dense<value_type>* result,
                                        array<char>& tmp) const = 0;
 
     virtual void compute_norm2_impl(
-        Dense<absolute_value_type>* result) const = 0;
+        matrix::Dense<absolute_value_type>* result) const = 0;
 
-    virtual void compute_norm2_impl(Dense<absolute_value_type>* result,
+    virtual void compute_norm2_impl(matrix::Dense<absolute_value_type>* result,
                                     array<char>& tmp) const = 0;
 
     virtual void compute_squared_norm2_impl(
-        Dense<absolute_value_type>* result) const = 0;
+        matrix::Dense<absolute_value_type>* result) const = 0;
 
-    virtual void compute_squared_norm2_impl(Dense<absolute_value_type>* result,
-                                            array<char>& tmp) const = 0;
+    virtual void compute_squared_norm2_impl(
+        matrix::Dense<absolute_value_type>* result, array<char>& tmp) const = 0;
 
     virtual void compute_norm1_impl(
-        Dense<absolute_value_type>* result) const = 0;
+        matrix::Dense<absolute_value_type>* result) const = 0;
 
-    virtual void compute_norm1_impl(Dense<absolute_value_type>* result,
+    virtual void compute_norm1_impl(matrix::Dense<absolute_value_type>* result,
                                     array<char>& tmp) const = 0;
 
 private:
@@ -804,7 +806,7 @@ void EnableMultiVector<ConcreteType>::scale_impl(any_const_dense_t alpha)
     std::visit(
         [this](auto alpha_v) {
             using alpha_type = std::decay_t<decltype(alpha_v)>;
-            if constexpr (std::is_same_v<alpha_type, Dense<value_type>>) {
+            if constexpr (std::is_same_v<alpha_type, matrix::Dense<value_type>>) {
                 this->scale_impl(alpha_v);
             } else {
                 GKO_NOT_IMPLEMENTED;
@@ -820,7 +822,7 @@ void EnableMultiVector<ConcreteType>::inv_scale_impl(any_const_dense_t alpha)
     std::visit(
         [this](auto alpha_v) {
             using alpha_type = std::decay_t<decltype(alpha_v)>;
-            if constexpr (std::is_same_v<alpha_type, Dense<value_type>>) {
+            if constexpr (std::is_same_v<alpha_type, matrix::Dense<value_type>>) {
                 this->inv_scale_impl(alpha_v);
             } else {
                 GKO_NOT_IMPLEMENTED;
@@ -929,5 +931,4 @@ void EnableMultiVector<ConcreteType>::compute_norm1_impl(MultiVector* result,
 }
 
 
-}  // namespace matrix
 }  // namespace gko
